+++ client/src/io/worxbend/gitea4s/api/IssuesApi.scala
@@ -2,8 +2,8 @@ package io.worxbend.gitea4s.api
 
 import io.worxbend.gitea4s.error.GiteaError
 import io.worxbend.gitea4s.http.IssueListParams
-import io.worxbend.gitea4s.model.{Comment, CreateIssue, EditIssue, Issue}
-import zio.IO
+import io.worxbend.gitea4s.model.{Comment, CreateIssue, EditIssue, Issue, Label}
+import zio.{Chunk, IO}
 import zio.stream.ZStream
 
 trait IssuesApi:
@@ -20,5 +20,15 @@ trait IssuesApi:
   def edit(owner: String, repo: String, index: Long, body: EditIssue): IO[GiteaError, Issue]
 
   def close(owner: String, repo: String, index: Long): IO[GiteaError, Issue]
+
+  def labels(owner: String, repo: String, index: Long): IO[GiteaError, Chunk[Label]]
+
+  def replaceLabels(owner: String, repo: String, index: Long, labels: Chunk[Long]): IO[GiteaError, Chunk[Label]]
+
+  def addLabels(owner: String, repo: String, index: Long, labels: Chunk[Long]): IO[GiteaError, Chunk[Label]]
+
+  def clearLabels(owner: String, repo: String, index: Long): IO[GiteaError, Unit]
+
+  def removeLabel(owner: String, repo: String, index: Long, id: Long): IO[GiteaError, Unit]
 
   def comment(owner: String, repo: String, index: Long, body: String): IO[GiteaError, Comment]